Become an AWS Kinesis Firehose client.
