+++ Gemfile
@@ -1,4 +1,8 @@
 source 'https://rubygems.org'
 
-# Specify your gem's dependencies in px-timeline-service-client.gemspec
+source 'https://gems.500px.net' do
+  gem "px-service-legacy-client"
+end
+
+# Specify your gem's dependencies in px-firehose-service-client.gemspec
 gemspec
