+++ lib/px-firehose-client.rb
@@ -1,0 +1,1 @@
+require 'px/service/firehose'
