--- lib/px-kinesis-client.rb
@@ -1,1 +1,0 @@
-require 'px/service/kinesis'
